Add GitHub Annotations feature to verify-bash.sh
This adds Annotations to the Job run's Summary and
in-line Annotations in the "Files Changed" tab

When env. var. GITHUB_ACTOR is present:

SHELLCHECK(1)'s json1 output format is used, JQ(1)
then outputs the GitHub API command:

    ::error file=<filename>,line=<line no>::<msg>

A simple GREP(1) is used to flip the return code,
i.e. "0" if no errors found and "123" when they
are. "123" is to keep in sync with the scripts
"normal", from XARGS(1) return code.

All of SHELLCHECK(1)'s severity levels:

    error, warning, info, style

Are reported as "error" for GitHub's API,
regardless of the severity (.comments[].level)