Hide our implementation of localtime_r in a namespace
Depending on time of day, MSYS2/MinGW fails when we redeclare
localtype_r (sometimes it fails because it's missing, sometimes it fails
because it's there).

Hiding our wrapper implementation inside a namespace should prevent
that. On POSIX systems we use alias to redirect cross::localtime_r to a
globally available localtime_r.