Cleanup GUS channel and IO handlers when inactive or shutdown
Allows for toggling the GUS device on-and-off with:

 1) config -set gus=false
 2) config -set gus=true

The "GUS" audio mixer channel is added or removed accordingly.

Regarding the ULTRASND= variable, the design of DOSBox allows
for its creation only during DOSBox startup, so the ULTRASND=
variable is not modified and left as-is in this commit.

However, the device itself will be present or absent, without
leaking or crashing.