Hide 'short global name' LGTM warning
LGTM warns about following short names:

- 'dos' in dos.cpp
- 'cpu' in cpu.cpp
- 'fpu' in fpu.cpp
- 'xga' in vga_xga.cpp
- 'vga' in vga.cpp

Honestly, we can do nothing about it - if we were to rename variable
name like 'cpu', it would cause havoc in the codebase and make it
impossible to sync with upstream.

Aside of cpu and fpu, these variables are limited to a single source
file only, so it's not a big deal.