Move environ pointer out of sdlmain
Also, replace undocumented 'environ' on MSVC (it's a macro) with
documented, deprecated '_environ'.

Remove ENVIRON_INCLUDED and ENVIRON_LINKED defines - they seem to be
pointless; perhaps a single checker is needed for future, but as for
now, all supported OSes have usable environ variable.

There are other options of accessing environment pointer, but they are
problematic - for longer explanation see:
https://wiki.sei.cmu.edu/confluence/display/c/ENV31-C.+Do+not+rely+on+an+environment+pointer+following+an+operation+that+may+invalidate+it