Limit the Disney Callback length to uint16_t
This matches the Mixer's API and the maximum possible length.

This 16-bit length is used to call DISNEY_PlayStereo(len, *, *),
which we update to match size-wise.  While we're at it, we also
constify both inbound arrays (which are only read), and tighten
up the loop that walk over 'len' range.