Manage MT-32 rendering and playback using the ReaderWriterCircularBuffer
The ring buffer makes the:
 - mixer-callback wait until a rendered buffer is available before playback
 - rendering thread wait until space is available in the ring before rendering

Previously, the playback position relative to rendering position along with the
thread state needed to be managed inside the mixer callback to prevent the
rendering thread from running away from the playback.  Now, with rendering
throttled by the ring buffer, we can eliminate the mutex, locking, and
conditional coordination in the callback.

We also can eliminate the equivalent management code in the rendering thread,
which was needed to coordinate the rendering position with playback position.

Rendering is now also performed in fixed-sized 2 KiB chunks, which is more
efficient than rendering to smaller and/or per-millisecond chunks.

The ring buffer also performs std::move() operations of the buffers to ensure
no copies are created.

Fixes issue #858 caused by an endless poll-wait loop experienced on some
Aarch64 SBCs.