Adjust allowed warning limits after enabling mt32emu
Move 'meson setup' to separate job step, so that build parameters are
visible directly on GitHub CI job page (makes it easier to see what's
happening).

Adjust CI for the fact, that mt32emu feature is now enabled by default.

Clang dynamic analyzer builds are failing - that's why the feature is
disabled in there. Other builds have feature enabled or disabled as
appropriate for the context.