Switch macOS release job to meson
This turns on macOS release job back on again. Link SDL2 dynamically, as
we have no other choice at the moment. We'll switch SDL2 to static
linking in a later commit.

Disable all other dependencies - they require brew libraries to be
installed on user's OS. We'll be re-enabling them one-by-one in later
commits.