Disable PPC64 CI workflow for now
Turns out this CI job is broken for some time now and Autoconf
buildsystem is building ARMv8 dynrec backend (?!) in here instead.

Upstream SVN r4424 started refactorization of cpu module, and it breaks
compilation of PPC and PPC64 backends; to be re-enabled once
refactorization is done or when we'll have time to fix it ourselves.