Avoid using direct memory lookup to read the terminal size
This method works for EGA/VGA, but does not work for Hercules or CGA any
more. Commit 9a1a0045deec574481af6ff82f8878e07ac81774 (SVN r4444)
changed the behaviour for old cards to improve detection of CGA mode by
some old games. Since that change, number of columns stored in BIOS is 0
for CGA, Hercules, and PCjr - and this broke cls command and dir
command.

At least for Hercules this change seems to make sense - it supported
only a single text mode, the same as MGA - which was 80x25.