Use a lookup table to convert 8-to-16 bit samples
See code and speed comparisons here:
 - GodBolt: https://godbolt.org/z/xvqe1K1en
 - Benchmarks: https://quick-bench.com/q/_nP_9Cc4SXVXvIfS6xYV9xdsH3M

When compiled with Clang 11.0 and -O2, the lookup table is:
 - 1.6-fold faster than the original bit-shifting code
 - 2.7-fold faster than the bias-corrected scaling code
 - 9.0-fold faster than floating-point conversion

When compiled with GCC 10.2 and -O2, the lookup table is:
 - 2.5-fold faster than the original bit-shifting code
 - 5.4-fold faster than the bias-corrected scaling code
 - 13-fold faster than floating-point conversion

Thanks to jmarsh for suggesting the lookup table and for pointing
out the original bit-shifting routine was negatively bias.

Thanks to hail-to-the-ryzen for the unbiased bit-shifting
comparison code and discussing this work on the vogons forums.

Co-authored by: hail-to-the-ryzen <hail-to-the-ryzen@vogons.org>
Co-authored by: jmarsh <jmarsh@vogons.org>
Co-authored by: kcgen <1557255+kcgen@users.noreply.github.com>