Process SDL's event queue at 200 Hz, the maximum PS/2 mouse rate
This change updates the frequency at which events are
processed from SDL's event queue to 200 Hz, which is the
maximum polling rate possible on original DOS hardware, and
thus provides the DOS application with a realistic number of
interrupts on original hardware.

This also fixes issue #1128, a regression discovered on macOS
when the 'Magnet' application is managing the DOSBox window.