Use DISPLAY prefix when logging display-specific details
The previous code simply used 'MAIN': while true in that
it indicates where the code lives (inside sdlmain.cpp), it
wasn't adding any value to those reading the logs in terms
of categorization.

The purpose of the prefix is to categorize the feature
being logged, so this change places all display-related
output into a 'DISPLAY' bucket for easier identification
(or grepping) for users.