Workaround GCC 7 and 8 that require experimental::filesystem
GCC 7 and 8 were released before the std::filesystem specification
was finalized in C++17, and use "experimental/filesystem" as
their include file and "std::experimental::filesystem" as the
namespace.

Newer GCC's and other compilers (Clang, MSVC) are currently
modern enough that no workarounds are needed for them.

Therefore, this commit uses a temporary "std_filesystem.h" that
normalizes these differences and provides a "std_fs" namespace.

The std_prefix was selected to explicitly indicate that the
file and namespace comes from std::, and not our own creation.

Unfortuantely we also cannot use the commonly suggested "fs::"
namespace shorthand due to the CPU core using an "fs" enum value.

When GCC 7 and 8 are eventually not supported, we can mass replace
"std_filesystem.h" and "std_fs::" with their C++17 equivalents,
<filesystem> and std::filesystem, respectively.