Revert clock doubling logic in VGA_SetClock
The concept of doubling the S3 clock for 8-bit VESA modes having the special
"line doubling" flag was introduced with the S3 work, which prevented flickering
in non-standard low-resolution modes used by build engine games.

This worked on all platforms except for the Pi4, where it causes video slowdowns
reported in #1297.  Therefore, it is being reverted, and a different approach
need to address the flickering in these modes.